#9 split up modules and better types
