+++ src/assert/index.ts
@@ -1,0 +1,2 @@
+export * from './is-leap-year';
+export * from './is-number';
