+++ src/random/index.ts
@@ -1,0 +1,11 @@
+export * from './boolean';
+export * from './date';
+export * from './date-day';
+export * from './date-month';
+export * from './date-year';
+export * from './from-array';
+export * from './geolocation';
+export * from './guid';
+export * from './hex-color';
+export * from './number';
+export * from './string';
